lib: picker3: java: minor adjustments

+++ picker/picker-color/src/main/java/androidx/picker3/app/SeslColorPickerDialog.java
@@ -99,7 +99,7 @@ public class SeslColorPickerDialog extends AlertDialog
                 if (mOnColorSetListener != null) {
                     if (mColorPicker.isUserInputValid() || mCurrentColor == null) {
                         mOnColorSetListener.onColorSet(mColorPicker.getRecentColorInfo()
-                                .getSelectedColor().intValue());
+                                .getSelectedColor());
                     } else {
                         mOnColorSetListener.onColorSet(mCurrentColor);
                     }
